Refactor to use a Root Handler.getHealth
