+++ src/api/metadata/handler.js
@@ -1,6 +1,18 @@
 'use strict'
 
+const { statusEnum } = require('@mojaloop/central-services-shared').HealthCheck.HealthCheckEnums
+const HealthCheck = require('@mojaloop/central-services-shared').HealthCheck.HealthCheck
+const Logger = require('@mojaloop/central-services-shared').Logger
+
+const packageJson = require('../../../package.json')
 const Config = require('../../lib/config')
+const {
+  getSubServiceHealthBroker
+} = require('../../lib/healthCheck/subServiceHealth.js')
+
+const healthCheck = new HealthCheck(packageJson, [
+  getSubServiceHealthBroker
+])
 
 /**
  * @module src/api/metadata/handler
@@ -30,63 +42,28 @@ const extractUrls = (request) => {
 }
 
 /**
- * @function health
- * @async
+ * @function getHealth
  *
- * @description This is the handler for /health endpoint
+ * @description Get the health of the service
  *
- * @param {object} request - the http request object
- * @param {object} h - the http response object
- *
- * @returns {object} - Returns the object containing the OK status and 200 status code
+ * @param {*} request - the Hapi request object
+ * @param {*} h - the Hapi handler object
  */
-exports.health = function (request, h) {
-  //TODO: Health check 
-
-
-
-  /* 
-
-  Health status design:
-
-  Design Goals:
-  - Clear HTTP Statuses (no need to inspect the response to know there are no issues)
-    - if response status is 200, then all good, everything is fine
-    - if status is 502, then API is good, but services maybe not. This may be the case when restarting the database, or when the api has started but not the database
-      502 is reserved for Bad Gateway, which I think kind of suits this case, but requires a bit of a loose interpretation of Gateway...
-    - if status is 503, then status is bad - the api won't be up and
-
-  example good status
-  GET /health 
-  200
-  {
-    status: 'OK',
-    uptime: number, (seconds),
-    upSince: string, (ISODate),
-    versionNumber: string, (from package.json)
-    services: [
-      {
-        name: 'kafka',
-        connectionStatus: 'OK',
-        latency: 12341 (ms)
-      },
-      {
-        name: 'mysql',
-        connectionStatus: 'OK',
-        latency: 12341 (ms)
-      }
-
-    ],
+const getHealth = async (request, h) => {
+  let responseBody
+  let responseCode = 200
+  try {
+    responseBody = await healthCheck.getHealth()
+  } catch (err) {
+    Logger.error(err.message)
   }
 
-  example bad status
-  503
+  if (!responseBody || responseBody.status !== statusEnum.OK) {
+    // Gateway Error
+    responseCode = 502
+  }
 
-  
-
-  */
-
-  return h.response({ status: 'OK' }).code(200)
+  return h.response(responseBody).code(responseCode)
 }
 
 /**
@@ -100,9 +77,14 @@ exports.health = function (request, h) {
  *
  * @returns {object} - Returns the object containing the hostname, registered URLs and 200 status code
  */
-exports.metadata = function (request, h) {
+const metadata (request, h) => {
   return h.response({
     directory: Config.HOSTNAME,
     urls: extractUrls(request)
   }).code(200)
+}
+
+module.exports = {
+  metadata,
+  getHealth
 }
